clean up, optimization and simplification
got rid of while loops, improved the functions to work better with many file holding files and older sfs containers with smaller chunk size.

It probably would be more practical to use xrange in for loops, but numbers are not huge and range in python 2 will work efficient enough. However in future porting to python3 will need no changes.