Add permanent markers (#1431)
* Early implementation of permanent markers

* Add _to_dictionary to MarkerBase

* Support for permanent markers via s.add_marker(marker,permanent=True)

* Add docstring for s.plot(plot_markers)

* Change s.add_marker docstring

To show permanent=True and plot_marker=True can be used at the same
time

* MarkerBase._to_dictionary: remove self_dict variable, use self.__dict__

* Check in s.add_marker to stop markers being used in different signals

* Add test for permanent markers

* s.add_marker, add check: if marker data dim work with signal dim

* Change Exception to ValueError in s.add_marker + corresponding tests

* Improve s.add_marker docstring

* Improve marker._to_dictionary: simplify getting the marker_properties

* Add plot_signal argument to s.add_marker

* Permanent markers: add missing argument

* MarkerBase: add plot_on_signal argument when initializing the class

* MarkerBase, _to_dictionary: save class as string, not type

* DictionaryTreeBrowser.as_dictionary will check for _to_dictionary

Will check if objects have the _to_dictionary function, and if they
do, get the dictionary there.
This for making markers savable HDF5 files when added to metadata

* Make Markers savable in dictionary in signal.metadata

* Permanent markers: markers saved in s.metadata.Markers will now plot

* Permanent markers: add plot_markers parameter + make plotting faster

* Permanent markers: fix failing tests + bug in adding marker twice

* Add more permanent marker tests

* Permanent markers: change marker names if several with same name

Start renaming counter at 1, instead of 0

* Permanent markers: fix replotting when adding multiple markers

* Permanent marker: add functionality to save and load

* Add name parameter to all markers

* Add test for saving and loading permanent markers

* Permanent markers: add test for saving and loading markers to HDF5

* Permanent marker: replot if s._plot.navigator_plot is None

* Permanent markers, IO: bind axes_manager to marker object in IO process

To make sure each markers gets access to the axes_manager in its
host signal. This to make sure markers which changes as function of
navigation positions works fine

* Permanent markers: add test for multidim navigation marker

* Add __deepcopy__ function to MarkerBase

Enables plotting of permanent markers with which changes
with navigation index, and s.deepcopy() with permanent markers
in the metadata.

* Add support for permanent markers in metadata for s.deepcopy()

* Minor improvement to s.add_marker docstring

* s.transpose: Add parser for markers saved in metadata

* Improve point marker docstring

* Improve s.add_marker docstring

* Improve s.add_marker docstring

* Add test for s.deepcopy with a permanent marker in the metadata

* Add try/expect to code loading markers

This to make sure that files with malformed markers dicts won't
stop the rest of the file from being loaded.

* Add IO test for bad marker type in HDF5 file

* Add user guide documentation for permanent markers

* s.transpose() will now remove the permanent markers from the metadata

When using s.transpose(), the navigation dimensions of the signal
will most likely change. And this might cause the markers to break.
So the safest is the simply remove them if transpose is used.

* PEP8 fixes for permanent marker implementation

* Revert retab in doc/user_guide/visualization.rst

* Fix MarkerBase.get_data_position not working when axes_manager=None

* Add __repr__ to markers

* Fix updating rectangle marker position

Was not working with the local matplotlib (1.5.2rc2).
Due to set_xdata and set_ydata not existing in the matplotlib
rectangle object.

* Make MarkerBase._to_dictionary more concise

Suggestion from @to266

* MarkerBase: improve _get_data_shape, and improve corresponding test

* Move function parsing raw metadata markers dict to drawing/marker.py

To avoid code duplication in s.__deepcopy and io.dict2signal

* Make markers_metadata_dict_to_markers more concise

Thanks to @to266

* Make s.add_marker and s._plot_permanent_markers more concise

Thanks to @to266

* Update markers section in user guide

* s.add_marker: remove plot_signal argument. Change plot_marker behaviour

The change in plot_marker behaviour might break some functions.
In the old version, s.add_marker would always call s.plot(), this has
been changed so if plot_marker=False, s.plot() will not be called.
This to allow for permanent markers to be added without opening
a plot.

* Add warning when using s.add_marker(plot_marker=False, permanent=False)

* Add test for loading marker in HDF5 with missing non-essential value

The test file has a marker where the y2 data is missing, however
it only needs the x1 and y1 data, so it should load. And work
since the missing non-essential values should be set to a
default value.

* Add test_dict2marker: tests all markers

* Improve tests for loading markers from HDF5 file

* User guide: add example on how to remove permanent marker

* Improve plotting speed when having many permanent markers

* Permanent markers: remove extra line

* Markers: when closing a figure, do not update after every marker removal

When closing a figure, all the markers are removed one by one.
Beforehand, the figure plot was updated after every single marker
removal. With many markers in a figure, this was very slow.
The updating after every marker removal was disabled

* Move plot function to MarkerBase, and add _plot_marker to marker types

* Move sanitize_dictionary to misc.test_utils

* Clean up HDF5 marker io test

* Update marker docstrings