Read navigation shape of .mib files from .hdr
Read navigation shape of .mib files from .hdr

Added a function to read the navigation shape and any eventual lineskips of a .mib dataset from the corresponding .hdr file. In case the data contains more frames than expected from the .hdr scan shape, the scan shape is compared to the number of frames to see if an integer number of frames are skipped per line (i.e. a lineskip>0 was used when acquiring the data). If a mismatch is detected, the function returns None rather than doing a best guess (the data is most likely incomplete, but it is not certain whether frames are missing inside the scan or if the scan was stopped before it was finished.

Also fixed some typos in various docstrings.

Created 270.enhancements.rst