app5 reader
Here is an app5 reader.  This works on the app5 files I get from our NanoMegas topspin system.  However, the formatting of the api might be totally different to how you prefer in hyperspy/rosettasciio.  So that might need discussion.  Anyway, this is a good place to start the discussion.